Re-architecture app
